reply function -v1

+++ src/main/java/find/cf/community/controller/PublishController.java
@@ -34,7 +34,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title",required = false) String title,
                             @RequestParam(value = "description",required = false) String description,
                             @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+                            @RequestParam(value = "id",required = false) Long id,
                             Model model,
                             HttpServletRequest request){
         model.addAttribute("title",title);
@@ -77,7 +77,7 @@ public class PublishController {
 
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Integer id
+    public String edit(@PathVariable("id") Long id
                          ,Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
